+++ frb_codegen/src/library/codegen/generator/api_dart/spec_generator/info.rs
@@ -212,7 +212,7 @@ impl<'a> ApiDartGeneratorInfoTrait for RustAutoOpaqueApiDartGenerator<'a> {
 
 impl<'a> ApiDartGeneratorInfoTrait for RustOpaqueApiDartGenerator<'a> {
     fn dart_api_type(&self) -> String {
-        rust_type_to_dart_type(&self.ir.inner.0, self.ir.brief_name)
+        self.ir.sanitized_type()
     }
 }
 
@@ -230,41 +230,5 @@ impl<'a> ApiDartGeneratorInfoTrait for StructRefApiDartGenerator<'a> {
                 .map(|it| it.to_code())
                 .join("\n"),
         )
-    }
-}
-
-fn rust_type_to_dart_type(rust: &str, brief_name: bool) -> String {
-    lazy_static! {
-        static ref OPAQUE_FILTER: Regex =Regex::new(r"((\bdyn|'static|\bDartSafe|\bAssertUnwindSafe|\+ (Send|Sync|UnwindSafe|RefUnwindSafe))\b)|([a-zA-Z0-9_]+::)").unwrap();
-        static ref OPAQUE_BRIEF_NAME_FILTER: Regex =Regex::new(r"(\bRwLock)\b").unwrap();
-    }
-
-    let mut rust = OPAQUE_FILTER.replace_all(rust, "").to_string();
-    if brief_name {
-        rust = OPAQUE_BRIEF_NAME_FILTER.replace_all(&rust, "").to_string();
-    }
-    rust.replace(char_not_alphanumeric, "_")
-        .to_case(Case::Pascal)
-}
-
-fn char_not_alphanumeric(c: char) -> bool {
-    !c.is_alphanumeric()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    pub fn test_rust_type_to_dart_type() {
-        assert_eq!(&rust_type_to_dart_type("SomeType", true), "SomeType");
-        assert_eq!(
-            &rust_type_to_dart_type(
-                "flutter_rust_bridge::for_generated::rust_async::RwLock<crate::api::simple::AnotherOpaqueType>",
-                true
-            ),
-            "AnotherOpaqueType"
-        );
-        assert_eq!(&rust_type_to_dart_type("flutter_rust_bridge::for_generated::rust_async::RwLock<(crate::api::simple::MyOpaqueType,crate::api::simple::AnotherOpaqueType,)>", true), "MyOpaqueTypeAnotherOpaqueType");
     }
 }
